fix(server): widen asset_command altitude to 32 bits
The server-side asset_command stored altitude as uint16_t while the wire
message (fss_message_asset_command) and the assets_assetcommand table both
carry 32 bits. An altitude command above 65535 read from the database was
silently truncated before being dispatched to the aircraft.

Widen the asset_command member, constructor parameter and getter to
uint32_t, and widen the ECPG read path (asset_command_s.altitude and the
db_asset_command_get host variable) to unsigned int so the full value
survives from the database to the wire. Add a regression test asserting an
altitude of 100000 round-trips through pack/decode and the session dispatch
path without truncation.

+++ src/fss-server.hpp
@@ -46,16 +46,16 @@ private:
     transport::fss_asset_command command{transport::fss_asset_command::asset_command_unknown};
     double latitude;
     double longitude;
-    uint16_t altitude;
+    uint32_t altitude;
 public:
     asset_command(uint64_t t_dbid, uint64_t t_timestamp, const std::string &t_cmd, double t_latitude,
-                  double t_longitude, uint16_t t_altitude);
+                  double t_longitude, uint32_t t_altitude);
     auto getDBId() -> uint64_t;
     auto getTimeStamp() -> uint64_t;
     auto getCommand() -> transport::fss_asset_command;
     auto getLatitude() -> double;
     auto getLongitude() -> double;
-    auto getAltitude() -> uint16_t;
+    auto getAltitude() -> uint32_t;
 };
 
 /* Pure-virtual database seam: lets ClientSession be unit-tested against
